Changed from GPT4
Make sure you run the command in the root directory of your project where the package.json file is located. This command will read the package.json file and install all the dependencies specified in the "dependencies" section.

After running the command, npm or Yarn will download and install the required dependencies into a node_modules folder in your project directory.

Note: If you previously ran npm install or yarn and are still experiencing issues, it's a good idea to delete the node_modules folder and run the installation command again to ensure a clean installation.